+++ app/frontend/src/component/mainpage/dm/Dm.tsx
@@ -1,31 +1,16 @@
-import { FC, useEffect, useRef, useState } from "react";
+import { FC, useEffect, useRef } from "react";
+import DmList from "./DmList";
 import "/src/scss/dm/Dm.scss";
-import { testDMList, DM } from "../../../dummydata/testDM";
-
-
-/*!
-  * @author yochoi
-  * @brief  msg가 20글자를 넘어가면 16글자까지만 출력하고
-  *         나머지는 ... 으로 대체하는 함수
-  */
-function msgFormatter(msg: string): string {
-  if (msg === null || msg === undefined) return (null);
-  if (msg.length >= 20) {
-    return (msg.slice(0, 17) + "...");
-  }
-  return (msg);
-}
 
 const Dm: FC<{isDmOpen: boolean}> = ({isDmOpen}): JSX.Element => {
 
   const dmRef = useRef<HTMLDivElement>(null);
-  const [dmList, setDmList] = useState<DM[]>(null);
 
   useEffect(() => {
     if (isDmOpen) {
-      dmRef.current.className = "dm-list in";
+      dmRef.current.className = "dm-container in";
     } else {
-      dmRef.current.className = "dm-list out";
+      dmRef.current.className = "dm-container out";
     }
   }, [isDmOpen]);
 
@@ -36,40 +21,15 @@ const Dm: FC<{isDmOpen: boolean}> = ({isDmOpen}): JSX.Element => {
   *        애니메이션이 실행됨
   */
   useEffect(() => {
-    dmRef.current.className = "dm-list";
-  }, []);
-
-  /*!
-  * @author yochoi
-  * @brief DM 리스트를 가져오는 useEffect
-  * @todo DM api 가 완성되면 easyFetch를 사용해서 수정해야함
-  */
-  useEffect(() => {
-    setDmList(testDMList);
+    dmRef.current.className = "dm-container";
   }, []);
 
   return (
-    <div className="dm-list" ref={dmRef}>
+    <div className="dm-container" ref={dmRef}>
       <div className="top-bar">
         <span>개인 메세지</span>
       </div>
-      <ul className="dm-list-ul">
-      {
-        dmList?.map((dm, idx) => {
-          return (
-            <li key={idx} className="dm-list-li">
-              <img
-                className="dm-list-li-avatar"
-                src={dm.target.avatar_url}
-                alt="dm target avatar"/>
-              <span className="dm-list-li-nick">{dm.target.nick}</span>
-              <span className="dm-list-li-last-msg">{msgFormatter(dm.lastMsg)}</span>
-              <span className="dm-list-li-last-msg-time">{dm.lastMsgTime}</span>
-            </li>
-          );
-        })
-      }
-      </ul>
+      <DmList />
     </div>
   )
 }
