+++ app/frontend/src/scss/dm/DmList.scss
@@ -1,0 +1,29 @@
+ul.dm-list {
+  display: flex;
+  flex-direction: column;
+  margin: 0;
+  padding: 0;
+  cursor: pointer;
+  li.dm-list-li {
+    display: grid;
+    grid-template-columns: 50px 70% auto;
+    padding: 10px;
+    list-style: none;
+    img.dm-list-li-avatar {
+      width: 50px;
+      height: 50px;
+      grid-row: 1 / 4;
+    }
+    span.dm-list-li-nick {
+      font-weight: bold;
+      grid-column: 2 / 3; 
+    }
+    span.dm-list-li-last-msg {
+      grid-row: 3 / 4;
+      grid-column: 2 / 3;
+    }
+    &:hover {
+      background-color: darkgray;
+    }
+  }
+}
